requirements.txt
fastapi
uvicorn
rembg
replicate
dotenv
pillow
requests
python-multipart
supabase
openai
numpy
tensorflow
keras
scikit-learn